+++ src/Ndax.Api/INdaxClient.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+using System.Threading.Tasks;
+using Ndax.Api.Models;
+
+namespace Ndax.Api
+{
+    public interface INdaxClient
+    {
+        Task<Dictionary<string, Instrument>> GetTickerAsync();
+    }
+}